backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-genai>=1.0.0
google-cloud-storage==2.14.0
google-api-python-client==2.108.0
firebase-admin>=6.2.0

